Fix: Delta upcasted to double bit-width and value downcasted after computation

+++ tsz-compress/src/v2/decode.rs
@@ -140,9 +140,9 @@ pub fn decode_i8<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i8>) -> Result<
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..10 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b111) as i8;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b111) as i16;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i16 + delta as i16) as i8;
+                    value = (value as i16 + delta) as i8;
                     output.push(value);
                 }
             }
@@ -162,9 +162,9 @@ pub fn decode_i8<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i8>) -> Result<
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..5 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111) as i8;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111) as i16;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i16 + delta as i16) as i8;
+                    value = (value as i16 + delta) as i8;
                     output.push(value);
                 }
             }
@@ -185,7 +185,7 @@ pub fn decode_i8<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i8>) -> Result<
                 for i in 0..4 {
                     let delta = ((word >> (shift - bit_width * i)) & 0b1111_1111) as i16;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i16 + delta as i16) as i8;
+                    value = (value as i16 + delta) as i8;
                     output.push(value);
                 }
             }
@@ -207,7 +207,7 @@ pub fn decode_i8<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i8>) -> Result<
                 for i in 0..3 {
                     let delta = ((word >> (shift - bit_width * i)) & 0b11_1111_1111) as i16;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i16 + delta as i16) as i8;
+                    value = (value as i16 + delta) as i8;
                     output.push(value);
                 }
             }
@@ -226,9 +226,9 @@ pub fn decode_i8<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i8>) -> Result<
                 let bit_width = 16;
                 let shift = 32 - padding - bit_width;
                 for i in 0..2 {
-                    let delta = (word >> (shift - bit_width * i)) as i8;
+                    let delta = (word >> (shift - bit_width * i)) as i16;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i16 + delta as i16) as i8;
+                    value = (value as i16 + delta) as i8;
                     output.push(value);
                 }
             }
@@ -247,9 +247,9 @@ pub fn decode_i8<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i8>) -> Result<
                 let bit_width = 32;
                 let shift = 32 - padding - bit_width;
                 for i in 0..1 {
-                    let delta = (word >> (shift - bit_width * i)) as i8;
+                    let delta = (word >> (shift - bit_width * i)) as i16;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i16 + delta as i16) as i8;
+                    value = (value as i16 + delta) as i8;
                     output.push(value);
                 }
             }
@@ -319,9 +319,9 @@ pub fn decode_i16<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i16>) -> Resul
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..10 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b111) as i16;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b111) as i32;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i32 + delta as i32) as i16;
+                    value = (value as i32 + delta) as i16;
                     output.push(value);
                 }
             }
@@ -341,9 +341,9 @@ pub fn decode_i16<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i16>) -> Resul
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..5 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111) as i16;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111) as i32;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i32 + delta as i32) as i16;
+                    value = (value as i32 + delta) as i16;
                     output.push(value);
                 }
             }
@@ -362,9 +362,9 @@ pub fn decode_i16<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i16>) -> Resul
                 let bit_width = 8;
                 let shift = 32 - padding - bit_width;
                 for i in 0..4 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b1111_1111) as i16;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b1111_1111) as i32;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i32 + delta as i32) as i16;
+                    value = (value as i32 + delta) as i16;
                     output.push(value);
                 }
             }
@@ -384,9 +384,9 @@ pub fn decode_i16<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i16>) -> Resul
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..3 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111_1111) as i16;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111_1111) as i32;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i32 + delta as i32) as i16;
+                    value = (value as i32 + delta) as i16;
                     output.push(value);
                 }
             }
@@ -405,9 +405,9 @@ pub fn decode_i16<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i16>) -> Resul
                 let bit_width = 16;
                 let shift = 32 - padding - bit_width;
                 for i in 0..2 {
-                    let delta = (word >> (shift - bit_width * i)) as i16;
+                    let delta = (word >> (shift - bit_width * i)) as i32;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i32 + delta as i32) as i16;
+                    value = (value as i32 + delta) as i16;
                     output.push(value);
                 }
             }
@@ -426,9 +426,9 @@ pub fn decode_i16<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i16>) -> Resul
                 let bit_width = 32;
                 let shift = 32 - padding - bit_width;
                 for i in 0..1 {
-                    let delta = (word >> (shift - bit_width * i)) as i16;
+                    let delta = (word >> (shift - bit_width * i)) as i32;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i32 + delta as i32) as i16;
+                    value = (value as i32 + delta) as i16;
                     output.push(value);
                 }
             }
@@ -510,9 +510,9 @@ pub fn decode_i32<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i32>) -> Resul
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..10 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b111) as i32;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b111) as i64;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i64 + delta as i64) as i32;
+                    value = (value as i64 + delta) as i32;
                     output.push(value);
                 }
             }
@@ -532,9 +532,9 @@ pub fn decode_i32<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i32>) -> Resul
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..5 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111) as i32;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111) as i64;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i64 + delta as i64) as i32;
+                    value = (value as i64 + delta) as i32;
                     output.push(value);
                 }
             }
@@ -553,9 +553,9 @@ pub fn decode_i32<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i32>) -> Resul
                 let bit_width = 8;
                 let shift = 32 - padding - bit_width;
                 for i in 0..4 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b1111_1111) as i32;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b1111_1111) as i64;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i64 + delta as i64) as i32;
+                    value = (value as i64 + delta) as i32;
                     output.push(value);
                 }
             }
@@ -576,9 +576,9 @@ pub fn decode_i32<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i32>) -> Resul
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..3 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111_1111) as i32;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111_1111) as i64;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i64 + delta as i64) as i32;
+                    value = (value as i64 + delta) as i32;
                     output.push(value);
                 }
             }
@@ -597,9 +597,9 @@ pub fn decode_i32<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i32>) -> Resul
                 let bit_width = 16;
                 let shift = 32 - padding - bit_width;
                 for i in 0..2 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0xffff) as i32;
+                    let delta = ((word >> (shift - bit_width * i)) & 0xffff) as i64;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i64 + delta as i64) as i32;
+                    value = (value as i64 + delta) as i32;
                     output.push(value);
                 }
             }
@@ -618,9 +618,9 @@ pub fn decode_i32<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i32>) -> Resul
                 let bit_width = 32;
                 let shift = 32 - padding - bit_width;
                 for i in 0..1 {
-                    let delta = (word >> (shift - bit_width * i)) as i32;
+                    let delta = (word >> (shift - bit_width * i)) as i64;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i64 + delta as i64) as i32;
+                    value = (value as i64 + delta) as i32;
                     output.push(value);
                 }
             }
@@ -726,9 +726,9 @@ pub fn decode_i64<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i64>) -> Resul
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..10 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b111) as i32;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b111) as i128;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i128 + delta as i128) as i64;
+                    value = (value as i128 + delta) as i64;
                     output.push(value);
                 }
             }
@@ -748,9 +748,9 @@ pub fn decode_i64<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i64>) -> Resul
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..5 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111) as i32;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111) as i128;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i128 + delta as i128) as i64;
+                    value = (value as i128 + delta) as i64;
                     output.push(value);
                 }
             }
@@ -769,9 +769,9 @@ pub fn decode_i64<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i64>) -> Resul
                 let bit_width = 8;
                 let shift = 32 - padding - bit_width;
                 for i in 0..4 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b1111_1111) as i32;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b1111_1111) as i128;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i128 + delta as i128) as i64;
+                    value = (value as i128 + delta) as i64;
                     output.push(value);
                 }
             }
@@ -792,9 +792,9 @@ pub fn decode_i64<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i64>) -> Resul
                 let shift = 32 - padding - bit_width;
                 word = word >> 2; // Nibble align
                 for i in 0..3 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111_1111) as i32;
+                    let delta = ((word >> (shift - bit_width * i)) & 0b11_1111_1111) as i128;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i128 + delta as i128) as i64;
+                    value = (value as i128 + delta) as i64;
                     output.push(value);
                 }
             }
@@ -813,9 +813,9 @@ pub fn decode_i64<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i64>) -> Resul
                 let bit_width = 16;
                 let shift = 32 - padding - bit_width;
                 for i in 0..2 {
-                    let delta = ((word >> (shift - bit_width * i)) & 0xffff) as i32;
+                    let delta = ((word >> (shift - bit_width * i)) & 0xffff) as i128;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i128 + delta as i128) as i64;
+                    value = (value as i128 + delta) as i64;
                     output.push(value);
                 }
             }
@@ -834,9 +834,9 @@ pub fn decode_i64<'it>(iter: &mut HalfIter<'it>, output: &mut Vec<i64>) -> Resul
                 let bit_width = 32;
                 let shift = 32 - padding - bit_width;
                 for i in 0..1 {
-                    let delta = (word >> (shift - bit_width * i)) as i32;
+                    let delta = (word >> (shift - bit_width * i)) as i128;
                     let delta = (delta >> 1) ^ -(delta & 1);
-                    value = (value as i128 + delta as i128) as i64;
+                    value = (value as i128 + delta) as i64;
                     output.push(value);
                 }
             }
